+++ src/components/Project/Resource/ResourceCards.tsx
@@ -21,6 +21,7 @@ import { DefaultThumbnail } from "@app/components/Project/Thumbnails/DefaultThum
 import MoreVertIcon from "@mui/icons-material/MoreVert";
 import { IncoreDialog } from "@app/components/IncoreDialog";
 import { VisualizationDialog } from "@app/components/Project/Resource/VisualizationDialog";
+import AddParentDatasetDialog from "@app/components/Project/Resource/AddParentDatasetDialog";
 
 import CheckIcon from "@mui/icons-material/Check";
 import { useNavigate } from "react-router-dom";
@@ -104,6 +105,15 @@ export const ResourceCards: React.FC<{
         }
         setOpenVisDialog(false);
     };
+    const handleOpenVisDialog = () => {
+        setOpenVisDialog(true);
+    };
+
+    // Update Source Dataset if they want to add a dataset to visualization
+    const [openAddParentDatasetDialog, setOpenAddParentDatasetDialog] = useState(false);
+    const handleCloseAddParentDatasetDialog = () => {
+        setOpenAddParentDatasetDialog(false);
+    };
 
     // batch selection
     const toggleSelection = (item: Hazard | Visualization | Dataset | Workflow) => {
@@ -141,6 +151,13 @@ export const ResourceCards: React.FC<{
                 open={openVisDialog}
                 onClose={handleCloseVisDialog}
                 onAddVisualization={handleAddVisualization}
+            />
+            <AddParentDatasetDialog
+                projectId={projectId}
+                open={openAddParentDatasetDialog}
+                onClose={handleCloseAddParentDatasetDialog}
+                resource={selectedItem}
+                handleOpenVisDialog={handleOpenVisDialog}
             />
             <IncoreDialog
                 open={openDeleteDialog}
@@ -226,7 +243,11 @@ export const ResourceCards: React.FC<{
                                                     e.stopPropagation();
                                                 }}
                                                 onClick={() => {
-                                                    setOpenVisDialog(true);
+                                                    if (isDatasetTable(resource) && !resource.sourceDataset) {
+                                                        setOpenAddParentDatasetDialog(true);
+                                                    } else {
+                                                        setOpenVisDialog(true);
+                                                    }
                                                 }}
                                             >
                                                 Add to Visualization